requirements.txt
<<<<<<< HEAD
adagio==0.2.4
aiofiles==23.2.1
aiosignal==1.3.1
alembic==1.13.1
altair==5.2.0
ansi2html==1.9.1
antlr4-python3-runtime==4.11.1
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
autoviz==0.1.804
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
bokeh==2.4.3
boto3==1.34.19
botocore==1.34.19
cachetools==5.5.2
catboost==1.2.2
category-encoders==2.6.3
certifi==2025.1.31
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.1
colorcet==3.0.1
colorlog==6.8.0
colour==0.1.5
comm==0.2.1
contourpy==1.2.0
cryptography==44.0.1
cycler==0.12.1
Cython==3.0.8
daal==2024.0.1
daal4py==2024.0.1
dacite==1.8.1
dash==2.14.2
dash-auth==2.0.0
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-cytoscape==0.3.0
dash-html-components==2.0.0
dash-table==5.0.0
dash-testing-stub==0.0.2
dask==2025.2.0
databricks-cli==0.18.0
databricks-sdk==0.44.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.7
distributed==2025.2.0
distro==1.9.0
docker==6.1.3
dtreeviz==2.2.2
dynaconf==3.2.10
emoji==2.9.0
entrypoints==0.4
et-xmlfile==1.1.0
evidently==0.4.40
exceptiongroup==1.2.0
executing==2.0.1
explainerdashboard==0.4.5
fairlearn==0.7.0
Faker==36.1.1
fastapi==0.109.0
fastjsonschema==2.19.1
ffmpy==0.3.1
filelock==3.13.1
Flask==2.2.3
flask-simplelogin==0.1.2
Flask-WTF==1.2.1
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fs==2.4.16
fsspec==2025.2.0
fugue==0.8.7
fugue-sql-antlr==0.2.0
future==0.18.3
fuzzywuzzy==0.18.0
gevent==23.9.1
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.38.0
gradio==3.50.2
gradio_client==0.6.1
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.20.1
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
holoviews==1.14.9
htmlmin==0.1.12
httpcore==1.0.2
httptools==0.6.4
httpx==0.26.0
huggingface-hub==0.20.2
hvplot==0.7.3
hyperopt==0.2.7
idna==3.6
ImageHash==4.3.1
imbalanced-learn==0.12.3
importlib-resources==6.1.1
importlib_metadata==8.5.0
iniconfig==2.0.0
interpret==0.5.0
interpret-core==0.5.0
ipykernel==6.28.0
ipython==8.20.0
ipywidgets==8.1.1
isoduration==20.11.0
iterative-telemetry==0.0.10
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-dash==0.4.2
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.4
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kaleido==0.2.1
kiwisolver==1.4.5
kmodes==0.12.2
lab==8.0
lightgbm==4.2.0
litestar==2.14.0
litestar-htmx==0.4.1
llvmlite==0.41.1
locket==1.0.0
lxml==5.1.0
m2cgen==0.10.0
Mako==1.3.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.6.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mlflow==2.20.2
mlflow-skinny==2.20.2
mlxtend==0.23.1
moto==4.2.13
mpltern==1.0.4
msgpack==1.0.7
msgspec==0.19.0
multidict==6.1.0
multimethod==1.10
multipart==1.2.1
multiprocess==0.70.15
mycolorpy==1.5.1
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.14.1
nbformat==5.9.2
nest-asyncio==1.5.9
networkx==3.2.1
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.24.4
oauthlib==3.2.2
openpyxl==3.1.2
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
optuna==3.5.0
optuna-integration==4.2.1
orjson==3.9.10
outcome==1.3.0.post0
overrides==7.4.0
oyaml==1.0
packaging==21.3
pandas==2.0.3
pandas-dq==1.29
pandocfilters==1.5.0
panel==0.14.4
param==1.13.0
parso==0.8.3
partd==1.4.1
patsy==0.5.6
percy==2.0.2
pexpect==4.9.0
phik==0.12.4
pillow==10.2.0
platformdirs==4.1.0
plotly==5.18.0
plotly-resampler==0.9.2
pluggy==1.3.0
pmdarima==2.0.4
polyfactory==2.19.0
prettytable==3.11.0
prometheus-client==0.19.0
prometheus-flask-exporter==0.23.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyamg==5.0.1
pyaml==23.12.0
pyarrow==14.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycaret==3.3.2
pycparser==2.21
pyct==0.5.0
pydantic==1.10.13
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
pynndescent==0.5.11
pyod==2.0.1
pyOpenSSL==25.0.0
pyparsing==3.1.1
PyPDF2==3.0.1
PySocks==1.7.1
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.6
python-ternary==1.0.8
pytz==2022.7.1
pyviz_comms==3.0.1
PyWavelets==1.5.0
pywedge==0.5.1.8
PyYAML==6.0.1
pyzmq==25.1.2
qpd==0.4.4
qtconsole==5.5.1
QtPy==2.4.1
querystring-parser==1.2.4
ray==2.9.0
referencing==0.32.1
regex==2023.12.25
requests==2.32.3
responses==0.24.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-click==1.8.6
rpds-py==0.17.1
rsa==4.9
s3transfer==0.10.0
SALib==1.4.7
schemdraw==0.15
scikit-base==0.5.2
scikit-learn==1.4.2
scikit-learn-intelex==2024.0.1
scikit-optimize==0.9.0
scikit-plot==0.3.7
scipy==1.10.1
seaborn==0.12.2
selenium==4.2.0
semantic-version==2.10.0
Send2Trash==1.8.2
shap==0.44.0
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sklearn-compat==0.1.3
sktime==0.26.0
slicer==0.0.7
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.51
sqlglot==20.8.0
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.35.1
statsforecast==1.5.0
statsmodels==0.14.1
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tbats==1.1.3
tbb==2021.11.0
tblib==3.0.0
tenacity==8.2.3
tensorboardX==2.6.2.2
terminado==0.18.0
ternary-diagram==3.1.0
textblob==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
triad==0.9.8
trio==0.24.0
trio-websocket==0.11.1
tsdownsample==0.1.2
tune-sklearn==0.5.0
txt2tags==3.9
typeguard==4.1.5
typer==0.15.1
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2025.1
ujson==5.10.0
umap-learn==0.5.5
uri-template==1.3.0
urllib3==1.26.20
urllib3-secure-extra==0.1.0
uuid6==2024.7.10
uvicorn==0.25.0
uvloop==0.21.0
visions==0.7.5
waitress==2.1.2
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==1.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==2.3.8
widgetsnbextension==4.0.9
wordcloud==1.9.3
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
wurlitzer==3.0.3
xgboost==1.6.2
xlrd==2.0.1
xmltodict==0.13.0
xxhash==3.4.1
ydata-profiling==4.6.0
yellowbrick==1.5
zict==3.0.0
zipp==3.21.0
zope.event==5.0
zope.interface==6.1
=======
# Main packages
pycaret==3.2.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.2

# Data handling
openpyxl>=3.0.10
imbalanced-learn>=0.10.1

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
shap>=0.41.0

# Utilities
python-dateutil>=2.8.2
joblib>=1.1.0
tqdm>=4.64.0

# For logging and error handling
rich>=12.0.0
>>>>>>> editando rutas relativas en real_world.py
